Refactor contact validation to further simplify
